Improve calling conventions, make hooks nest
If \diglottrue is set, then ensure that the diglot routines are
actually being called (previously \BodyColumns had to be properly
specified).
Also, make the hooks call in a nestable order.

I'm still not sure if it makes more sense to have hooks for e.g.
\addL and \add function both function or if a hook is specified
for \addL it should override the hook for \add.
David