restore \pr@tect, more space for wide callers
Don't know why restoring protect  is needed, (it ought to be a local change)
or what's triggering the NBSP error, but it seems like restoring it might help.

0.7ex is too small if the caller is as wide as @ (rare). Increase to 1ex ought
to be plenty, and not too big. An alternative solution is to find where the
caller is used in the notes and box /that/ to a fixed size, rather than the
definition of the caller. To be investigated.