Improved support for pages with no verses.
Previously, if a glossary or other book was started that had content but no
title, the 'final' chapter and verse would be marked in the header of all following pages.
If a title was specified, then title-style headers/footers would be used for all pages.
Also, the RHtitle.... positions were not included, presumably due to a programming error.

Patch introduces new set of running header and footer positions, for odd and
even pages without verses (mnemnonic: noV[erses]):

\RHnoVoddleft, \RHnoVoddcenter, \RHnoVoddleft
\RHnoVevenleft, \RHnoVevencenter, \RHnoVevenleft

and RF equivalents for footer.

So far, diglots probably don't activate these headers, a (different) approach
that was tried didn't work and has been commented out.