Update documentation, fail to resolve wandering pictures
post-paragraph pics:
If at the end of \par (which is where the post-paragraph pics get added to the
stream) you add a box (picture, but \hbox{x} or \hrule will do too) then
NORMALLY, it just appears in the output, as expected. BUT if the paragraph
happens to cross the page then the continuation page is entirely blank. e.g.
picture inserted (piclist) in v18, with a par that started in v16 on the
previous page -> blank page where v17&18 etc.  ought to be (page numbers also
missing).  If I force a par in v17, then all works properly. In diglot, the
image 'merely' ends up at the start of column left, no matter which side it should
be set on.