Image alignment changes and documentation
*i*nner and *o*uter are now valid alternaives to *l*eft and *r*ight for all
picture placement / alignment options.

Just don't use ti / to with diglotInnerOuter, or total confusion will reign.

Page-edge alignment of images is a complete pain. The cropmark adding routine
includes \vss above and below the final box, and applies \topm@rgin and
\botomm@rgin above and below. Thus it does NOT actually expect an actual full page.

Also, for some reason there seems to be a miscalculation of paper size - (by TeX or
by evince??) the top of the paper seems to extend a couple of points higher
than TeX believes, but the bottom seems correct. I think I've finally got it working,
but added a (user-controlled) fudge-factor.