Add slurp-mode processing of piclists, with error reporting.
\picslurpfalse : traditional piclist rules apply

\picslurptrue :
* Read whole piclist, generating a unique macro for the picture
specification based on the reference and the image number for that reference.
Also make a unique macro containing a
human-friendly version of the reference (e.g. GEN 1.2(2) for the second picture
on GEN 1.2), and put that macro on a list.
* Increment the count of defined pictures.

* When a picture is used, blank the macro with the human-friendly version of the list, and increment the count of used pictures.

* If the picture counts don't match at the end of the document, display the
list. Only those macros that haven't been displayed will generate any output.
This is all probably quite memory hungry, hence keeping the traditional version available.