Work around some odd futurelet/expansion behaviour (#207)
The effects of active characters in captions from a piclist seems unusual, there seems to be some
double expansion, or something else unexpected. use of a unique \next value which is reset to \relax
after each use seems to solve this. It's not at all clear to me why this is so.

Also put customised header back on POT file.