Support for categories in footnotes (and maybe other situations?)
Note that the category must be set before character styles are imposed,
otherwise the nesting rules will remove the category setting. This is in
accordance with the USFM use here:
http://ubs-icap.org/chm/usfm/3.0/notes_study/categories.html#cat-cat

That document specifies this is a feature for \ef  and \esb only. As
implemented here it applies to all note styles.

E.g.  usfm:
 \f + \cat people\cat*\fr 7:11 \ft \fq Aaron\fq was of the clan of Levi.\f*

Direct stylesheet markup is functioning:

\Marker cat:people|fq
\Color x7FFF7F