Use catcode juggling to find end-mark of attribute in non-milestone marker
still need to *do* something with them, but capturing them seems to work.

This is 3rd of 4 alternatives tried so far. Alternatives 1, 2 and 4 did not work
One not yet tried would be to put the text in a box, as with headings, etc.
and then unhbox it. For some reason alternative 1 - \def\foo\bgroup} does
not work, possibly because of intervening macros.