start on getting char-styles with attributes to be sensible
Text in attribute-possible char styles are pushed into
\hbox0, and either unhboxed or if attributes were
found, processed by \csname complex-\style\endcsname if that
exists.

Helper macros will need defining e.g. \style@ttr{attributename}
could spit out the attribute in the correct font for the ending
style and that attribute.
It should also check publishable to see if the attribute is to
be displayed or not.

See discussion in Milestones.md