Shell script to update Windows XeTeX binaries
This is a copy of what I have used in the past to update XeTeX
in Paratext for Windows in a mostly manual process. In the future,
this script will automate the process of updating XeTeX for use
with PTXprint on Windows. PTXprint on Linux will continue to use
XeTeX from the system packages.