Automatically update TeX tree with newer TeX Live files
In a previous commit of this file (before the renaming)
the initial package list was stated to come from http://w32tex.org/
in the section Minimal installation.

While this is correct, there were other packages added to the list
(dvipdfm-w32 and xetex-w32) to get PDFs from XeTeX.

The current list in this file is a trimmed down version of the
first list. The contents of each package was compared to what
was shipped with Paratext for Windows.