Update to TeX Live 2020 (on Windows)
You can check which version PTXprint is using (either it's own
version or the version in Paratext) by looking at the XeTeX log.
The first line will have TeX Live 2019 if the Paratext version is
used, if TeX Live 2020 is in the first line, then the internal
version is being used.

These version numbers are current as of now, in the future PTXprint
but maybe not Paratext will be updated to TeX Live 2021.