Piclist entries referring to paragraph should be marked with equal sign
The code needs to behave differently to define a trigger after a paragraph break ('preverse' is not accurate).
However,  '-' is too common as a hyphen in key terms, which are now valid anchor points. Thus the never-working
chapter.verse-paragraph format is now changed to chapter.verse=paragraph (or keyterm=paragraph), which actually works

GEN 1.2=3 refers to the third paragraph  of Gen 1.2
GLO Kiriath-Baal=2 is the 2nd paragaph on the entry for Kiriath-Baal