Cope with milestones straight after verse numbers
This patch removes the treatment of numerical verse identifiers as numbers, treating them
instead as a string. This is something which has been in the code since 2007 or before,
but it does not appear at present to be required anywhere. (It would be if the verse number
were used anywhere in a calculation).