Remove diglot@backingup, its fundamentally flawed.
No one can guarantee there's not a breakpoint and then just glue before the
(critical) final end-of-trial penalty. The page-breaker will discard
glue and penalties after a break. Therefore the code mustn't use the page
breaker to split galley / galley excess.  This was done to preserve the
penalty at the break point, but that gain in no way outweighs the damage.

The monoglot backing up routine doesn't need to worry about this issue as it
wants to lose the tail end of the current galley in what comes next.

This patch solves the above problem, using vsplits instead. Extra
checks are needed, as vsplit will *always* split off a line if there's
no additionally inserted break-point at the start of the box.

Also, it was necessary to make to be based on available space, not trialheight,
as vsplit uses a simplified algorithm that does not try to include any
calculations for inserts/notes.

Related to these changes, the vsize-related logic in diglot@anytrial also
needed updating to make use of the most recent availheight calculations each
cycle, and merge in additional chunks of text as appropriate.

For future study: It may be appropriate to be overly generous with
vsize on the trial runs, reducing the need for joints and repeated
output cycles.