Fix adjust list triggering based on milestones/keywords
Adjust lists can now trigger based on milestones or keywords much like
piclists.
Not having a . in the "chapter.verse" field of the adjust list will no
longer cause a parse error, but a mistake like "chapter,verse" will simply
stop processing in the same way that GFN instead of GEN would.

It used to be possible to use an all-caps milestone of XX.YYY form.
There was a processing error where:
```
\id XXA
...
\p bla bla \zfiga|A.ABRAHAM\*
\p Second paragraph
```
And an adjust list entry of
```
XXA A.ABRAHAM +1
```

Would not trigger for the line with the milestone in and treated the
second paragraph as the 1st paragraph of "A.ABRAHAM".
Users using this old work-around will need to adjust their paragraph
offsets by +1.

Also corrected handling of paragraph-onset trigger points such as:
```
\p \zfiga|Foo\* Text text text
```
This may affect some picture placement.