Endnote woes
Paragraphed endnotes are expanded to the hsize of the column in which they are expanded.
Non-paragraphed notes are paragraphed at definition, which causes problems. Ideally they
should be set to the hsize of the place where the endnotes are going to be
expanded.  Current hsize is probably less problematic that textwidth.