Improve docuentation, Add syleshet control \NonJustifiedFill
\NonJustifiedFill is the maximum amount of fill that can be applied to a right
(or left) aligned paragraph, measured as a function of the column width.  It
makes to increase this for poetry which might have user-specified breaks

// now increases linepenalty (for that paragraph) and negates it at the breakpoint.
Assuming above NonJustifiedFill is long enough, a line with // will only break if it has to,
but if it has to it will break there. Older version would break at // even when the
break point wasn't necessary.