textborder on whole diglot page.
The "easy option" for textborders is to wrap the whole page.

Single/multiple column textborders are significantly harder, as chunks of text
are added to the page by row, and it will be necessary to determine where the
start-point / height of scripture-containing chunk(s) start and stop on a page,
and "over-print" them with the border.