Di/Polyglot style(sheet) switching without polyglot columns
Introducing `\monoglotcolumn R` (or L, A, B, etc). Without activating the
column-switching code, this activates the diglot sense of `\p` etc,  specifying
which column's settings should be used.  Thus side-dependant settings can be
used in book introductions, etc.
The argument is a single capital letter, so that it can occur in a USFM file.