Remove not running empty ptxfile before main job
If there is front matter this is run before the main files, as a ptxfile. But if
there is no font matter, ptxfile is called with an empty filename. It would seem
natural to comment this out. But this causes all kinds of problems for diglots,
so we revert that change.