Fix noinkinmargin positioning
With no vfil to  push it down, the noinkinmargin footer box was drifting up to
the bottom of the text. We may need to check that the footer box has the
requisite space in it at the bottom to fit appropriately.

Updated test results are included in this commit. Things look good except:
  my_judson/ptxprint-better-ROMmy_judson.pdf
which has poor footnotes. This could imply some other footnote line spacing
issue more generally.