Even better hook behaviour??
A change to the behaviour in terms of complex hooks that now if multiple hooks
*could* apply to a certain point, only the most specific hook is actually
applied.
This may require some customisations to be re-written. The motivation for this change is
that if a certain behaviour is not wanted in a certain location, but required generally,
it is hard to *undo* certain changes, which is effectively what the old code wanted.

Note that specifying a hook with and empty action deletes any previously set version
from memory. To specify a do-nothing hook, e.g. to use the override functionality
mentioned above, an action of `\relax` should be passed.