Remove slashes from \k...\k* reference (related to #766)
The value of a key term is turned into a reference mark. That reference mark may
be used as an anchor for an image or a trigger, but it also crops up in the \parlocs
file in some situations. If a glossary entry is \k The \+nd Lord\+nd*\k*, then
it is thus not enough to remove spaces to make a usable anchor point; slashes must
also be removed.  (The above reference now becomes `k.The+ndLord+nd*`).