Fix issue with regurgitated char styles (#766)
k processing and maybe some other things end up with 'regurgitating' markers.
This can add spaces into a sequence:
 `\+nd Lord \+nd*`  becomes `\+nd Lord \+nd *`
 which breaks end-of-char style detection.

Regurgitated code occurs in known locations (\k processing) and can be searched
for.  It is only believed to only be relevant when the new marker is equal to
the current marker.