Xobjects continue to cause problems
New \holdXobjectstrue/false to try moving them into the top of the shipout.
For some reason, despite the logs insisting it's the other way round,
the use is still ending up in the xdv before the definition, either with objects
held or not.

Xform objects still crop the edges of their content.

Therefore, DON'T set \XformOrnamentstrue.