Quick and dirty hack to fix secondary styling
This leaves unused files being generated, and the probably confusing status-quo that if any USFM file is loaded
without a preceding \zglo|L\* it will (continue to) be immune to styling.

This includes the FRT book, for which users  have to specify which styling-set they want to apply
to which bit (via zglo). (starting the FRT book with zglo in force leads to
some strange formatting glitches; the exact cause is unknown).