Implement zvar access for milestone /  char style attributes
If a milestone attribute (exactly) matches the pattern "zvar::variablename"
then the named variable will be returned by get@ttribute.
This is to get around the issue that it is impossible to include one milestone
inside another. It is NOT intended to allow generic mid-string replacement.
People wanting that level of cunning complexity should use
`\csname ms:zvar=variablename\endcsname`