Feature maths: extend should probably multiply.
When p^95 is applied to a character style which sets its own extend, the result
should probably be that the final extend is 0.95 of the character style's extend.
The current solution assumes this is a *general rule* and that any extend
should multiply a nested extend within it.  I.e. a charstyle with an attribute
of extend=1.1 in a paragraph style with an attribute of extend=0.9 should
result in a final value of 0.99.

This is in contrast with the old behaviour where the character style would
end up with an extend=1.1.

It's just occurred to me this may not be the desired result, and that the
p^95 value could have been applied to each values. This would, however, make the
p^95 behave in a different manner to the behaviour of setting extend=0.95 on the
paragraph.