Bodge a partial fix for separated book titles
The old code tests \PageFullFactor too many times:
 * On switching to single column from an end of book
 * On switching to double columns.
The result of this is that book titles were often on an early page, but
although there was space for the text to start, it did not as the page
*now* did not meet that criteria.

This commit replaces the second check with >4lines of text, as an approximate
minimum. It does not guarantee that there will *actually* be sufficient space
on the page, and has no knowledge about there being space for section headings,
etc. on the page.

There are at least 3 better solutions:
1. Calculate separate PageFullFactors for titles and body / intro text.
2. Hold on to the book title before determining whether there's space for it to go on the
current page or if there should be a page-break.
3. Somehow use penalties to pull the heading onto the next page.