Make ifFinalNotesDown work properly
ifFinalNotesDown should make footnotes on the last page of a document appear at the bottom
of the page, rather than floating up. The way it does this is to insert
vertical stretch in the page. That stretch gets ignored if that is in an inner box,
which it used to be in some circumstances. Unwrap those boxes ifFinalNotesDown