Update to TeX Live 2023 (on Windows)
Previously, the TeX tree came from the W32TeX project https://www.w32tex.org/
which was a repackaging of Tex Live. However this site has not been updated
recently so the new TeX tree comes from the main TeX Live project https://tug.org/texlive/

The previous commit moved files in the TeX tree to new locations as W32TeX
and the main TeX Live project have slightly different layouts. The biggest changes
are the location of the executables (bin/windows instead of bin) and xetex.fmt
has moved (texmf-var/web2c/xetex instead of share/texmf-dist/web2c/xetex), thus
creating a new directory in the xetex folder (that is, texmf-var).

The file ls-R was copied to both texmf trees as it seems to be needed in texmf-var.
The file was also added to texmf-dist for completeness.