Make zfiga work on the cover
The cover is now treated as a special book (id: COV) as far as triggers, etc. are concerned.
piclist entries for COV are now included in the job piclist.
This means that e.g. \zfiga|logo\* can pull in a publisher's logo, for
positioning and sizing via the normal piclist editor.