Track down an odd font in a RTL (nastiliq) project
In drawing the ISBN, the code tries to make the top text (ISBN 123-45-6789-0)
exactly match the bar code width, by adjusting its size.  When the bug
manifests, this fine-font adjustment causes XeTeX to crash.
This is very odd, and that sort of font change is something we're doing all the
time, e.g. with the p^99 code.

A work-around is to use the same font-size as the below-barcode text. That can be
triggered by setting the new boolean, \noScaleAboveISBNtexttrue