First attempt at #903
Currently, it's using x-edgeadjust="0.1" in figures/piclist, and
`\FirstLineIndent 0.1` for sidebars.  It is thus a bit of an inconsistent
mishmash, and needs improving.
Units are `\IndentUnit`s, like margins, 1st line indent, and positive numbers
indicate 'inward' towards the text body. I.e a negative value moves it into margin area.

It does not affect the position of centered images. Nor images in inserts (tl,
etc), as those are centred.