Warn when overwriting non-stylesheet macros with stylesheet
The technique is not trying to be fully accurate - stylesheet definitions that
have been overridden with a special macro will not flag an error, but if e.g.
there was a macro \pb defined (normally) in the macros (as there was), and that was
then overridden by the stylesheet, that would be flagged.