Cope with mt1 on line after an id
The 'grab' code that allows crash-free processing of \id and \h when there are insufficient line-breaks (sfm on these important
lines), and also multiple-line id / h fields, has the side-effect of reading the control sequence with whatever cat-codes are
in force during the grab. \id sets digits to be processed as numbers. (Motivation for this is currently unknown).
This patch ensures that after a new line, digits are treated as letters once more, so that e.g. an \mt1 immediately
after an \id line is not treated as `\mt 1`.