Implement @diglot@pagebreak
It has been noticed that there is a merge issue on the 'doc' method, with intro
paragraphs including bodytext paragraph marks `\p`, triggering the 'to body'
transition once per intro paragraph.  'Scored' does not exhibit this behaviour.

This patch does not clear that problem, but if the merged file is correct, then enabling
PBOnBody now results in the correct paragraph break behaviour for diglots.