Avoid lost last partial-page of SFM input.
Including a PDF file after the last USFM file, when `\ifendbooknoeject` was true
could result in the (total) loss of the last partial page of SFM. The page is stored in
the `\partial`, and if there is no non-discardable vertical material after that, no further
output routine would be called.
Two changes have been made:
1. partial is always emptied by `\end`.
2. a new boolean, `\iffinalbooknoeject`.  If the new boolean is true, then any residual
text is output by (1) above, as the very last page of output. This is normally not desired,
however, and the default (false) value sets overrides `\ifendbooknoeject` on the final book,
*after* the bookend-final trigger (used for colophons) has triggered.