Make custom merge config files work
merge-option.cfg now operational (for scored methods). This allows control over
things like whether section headers are in a chunk with the following verse or not.
If no configuration file is found, one will be written. Note that as the config
files are persistent, damage to one will *also* be persistent. Delete the file to restore
things to normal.