Figure inclusion / triggers: obey paragraph numbering for start-of-paragraph figures'
This will change some layouts.  Bump TeXModel._ptxversion to 4 (It should have
been 3 for quite a while).

Reminder 1:
\p \k Temple\k* in GLO is referenced by `GLO k.Temple` in piclist.
The 2nd paragraph of that is `k.Temple=2` (which is triggered at the start of
the pargraph).
k.Temple=1 represents the first paragraph, and *may* get triggered, if
`\k Temple\k*` is the first printable thing in the paragraph, but that should
not be relied upon. Normally it is an error to use `k.Temple=1`
Chapter.Verse and zfiga anchors obey the same pattern. (i.e it's normally a
mistake to use =1)

Reminder 2:
pgpos="h" means 'here',
where that might be just above the verse number if the verse is in the middle
of a paragraph.
pgpos="pc" means centred, after the specified paragraph has ended.
pgpos="pc2" means after the *next* paragraph has ended.

Reminder 3:
There's an important but subtle difference between `GEN 1:2=2` (or `k.Temple=2`) ... pgpos="pc"`
and
`"GEN 1:2 ...  pgpos="pc2"` The former will only display the picture if there is a
paragraph break in the middle of Gen 1:2 (or the description of Temple). The latter
will display the picture after the following paragraph, no matter what the verse number /
key term it is.