Solve the immediate problem of galleys ending in skips which cause EEKs
A galley-final skip with a short trial height will break before the skip and then discard both skip and
galley-end penalty, resulting in lost synch, etc. Remove skip before running
trial, and (if space) put it back after.

A bit of foundation for a refactor that's probably needed for figures to work better -
collect defines a per-column \wholeX in case the page needs reflowing, (which then gets split into partialX)