Reconfigure @parlen reporting to help python identify pars
Immediately preceeding format was a bit inconsistent:
{Current ref}{old paragraph number}{paragraph lines}{marker type}{current paragraph number}
(Originally, instead of {current ref} for parameter #1, it had been {old ref})

New format:
{Current ref}{current paragraph number}{paragraph lines}{marker type}{old ref}{old paragraph number}
Where 'old' means the reference/paragraph when the paragraph first started.
For a paragraph which is \p \v 3 ... the current and new values are the same,
but for  the situation \v 2 text \p text  \p text \v 3, then the final
paragraph will start as v2 para 3, and end with v3 para 1.  "\p The \k temple\k*"
is similar.

Also fix logic so that when the a milestone isn't unique, restoring the current
ref. (temporarily set for triggers) occurs correctly.