Don't use xdef when storing periphs
Storing a periph for future use is fraught with issues. Some methods (those
that tokenize and then detokenize the input) result in the automatic addition of
a space after all command-sequences, even those like \nd ... \nd* where the
presence or absence of a space is significant. Thus code makes \ a normal letter and
tries to store the code as-is.

Previous to this patch, lines were added using \xdef. As (normal) macros were
invisible, they were not expanded. However active characters in the periph content
*were* expanded, which occasionally caused problems. This patch replaces the xdefs with a suitable
if unsightly heap of expandafters, so that only the macros involved in the capture are
expanded and nothing else.