Add code (inactive) to differentiate between structural and non-structural char markers in notes
USFM3.1 is somewhat inconsistent in its treatment of 'character markers', in
that it treats most markers as automatically nesting and certain markers as
'structural'. There doesn't seem to be stylesheet support for differentiating
these marker types, but the suggestion is that the code needs to handle mixed
data, such as:
 \f + \fr 3:14 \ft This \it is\it* a test\f*
Where \ft is treated as implicitly closing \fr (structural), but \it nests.

This commit adds code to treat (some?) structural tags correctly, while other
markers are treated as nest-forming.

The new code is not yet active, as a test suite is needed. Instead, the
previous situation us preserved, where the boolean \implicitcl@setrue is set
when in a footnote, causing processing to revert to the usfm3 state, where all
non-plus markers are treated as causing the implicit closure of previous markers.