Replace archive-specific font lookup location
To use runtex.sh there needs to be an entry:
    <dir prefix="cwd">fonts</dir>
in the archived version of fonts.conf. (or fonts need to go somewhere else when
the archive is built)

This change was added once before (march 28), and then got removed by another
patch (2nd April), I guess accidentally as part of a conflict resolution.
Hence it stopped working.