Solve character-styles in tables issue
Like footnote parts, table cells are *structural* with the opening of one
implicitly closeing of the previous structural part.
* Semi-unify structural elements to define
\csname STR@CT-marker\endcsname for both note and table markers. So far these are only
counted as structural  when used in appropriate environments (\ifdoingtable, \ifinnote).
I don't know if that restriction is necessary, but it seems sensible.

A new macro (@settablemark) sets up the table cell marker and *also* classes the marker
as structural. This paves the way for programatic generation of table cell markers,
if that were desired.