Use the right penalties.
When the pagebuilder is deciding whether to break a note, there should be a interfootnotelinepenalty,
but when control is passed to our code and we're rebuilding, we need penalties of >=10k, so that the
notes don't wander.  In case they need identifying, (debugging purposes), I've made this a variable,
earlyinternotepen@lty. The @ sign is a warning that users shouldn't touch it.