fix #1091
Tricky one to track down.... The end-this-layout code was (reasonably enough)
getting called by the transition from intro text to verse text. However,
the end-this-layout code was triggering the end-of-page content, which includes
the commands \beginL\savepos, and it seems that \beginL triggers \everypar.
The \everypar code (remaining from the intro passage) noticed there was a
chapter waiting to be typeset, and output the chapter number in the end-of-page
content. It is *NOT* clear to me why this is a rare bug,
possibly it is because there was a normal (not heading) paragraph as the first chapter
content.

The attached solution ensures that everypar is blank when calling the
layout-switching code, and (to catch other routes) also in line triggering the
end-of-page content.