Investigate and somehow solve adjustlist-caused maxdimen-wide pages
If an adjustlist entry affected the final note on the page, the classic 'blank
page except a few words' symptom of a failed footnote processing was seen.
(Output tracing in this case will show maximen-wide footnote boxes). The
triggering line is not known, and is presumed to be unexpected
onward-expansion, such as might be triggered by \box3\if....