Protect figure borders from peripheral interaction, semi-solve map borders.
When processing a periphery book, the end-of-file marks the end of the periphery,
and needs to trigger various processing, including ending several groups.

This means an end-of-file (such as from inputting an ornament, or \scantoks)
will thus cause untold confusion if it happens while a figure is being processed.

Very similar (but subtly different) group-related chaos occurs if (while in a
periphery) the normal \par code is encountered while processing the
end-of-this-par code used to include full-page figures,

Corrected catcodes also means map borders now seem to be almost-working.
For some reason the font outline was not being applied.