Introduce per-column catergories, make textborder one
Until now, all 'category' parameters (sidebar-specific parameters, which
includes borders) applied to all columns of a diglot equally.
This is considered appropriate styling, normally, and allowed either column
to trigger the sidebars or the cover.

textborders are a bit more messy, however.
It ought to be expected that they could be configured globally or per-column,
and the first part of this change-set was to allow that latter step to happen
correctly (breaking the previous model for textborders).
However, the user-interface does not allow for the configuration of
publication-wide style settings  (all styles are linked to a given project).

This meant that setting up a 'global' text border and then defining
\TextBorderColumns{R} (only put a border around the secondary text), could
disable the border until the user realised that the border needed to be
enabled and all its settings defined on the secondary project. This was unintuitive.

As a work-around, code has been added so that if L (primary) has an enabled
textborder, but a secondary column results in 'nonpublishable', the code should now:
1) override any 'nonpublishable'  status on the secondary column.
2) look up the settings for the textborder with a priority order of secondary, primary, defaults.

Thus, if the secondary stylesheet has no non-default settings the primary stylesheet values
are used for a textborder on either side.